Make SISeconds' integer literal initializer take Int instead of Double

+++ Sources/Time/1-Agnostic Types/SISeconds.swift
@@ -100,7 +100,7 @@ extension SISeconds: ExpressibleByIntegerLiteral, ExpressibleByFloatLiteral {
         self.init(value)
     }
     
-    public init(integerLiteral value: Double) {
+    public init(integerLiteral value: Int) {
         self.init(value)
     }
 }
